WIP add view argument to perform method
