+++ Wire-iOS/Sources/UserInterface/Conversation/Content/ConversationContentViewController+ConversationMessageCellDelegate.swift
@@ -19,10 +19,24 @@
 import Foundation
 
 extension ConversationContentViewController: ConversationMessageCellDelegate {
-    public func perform(action: MessageAction, for message: ZMConversationMessage!) {
-        guard let cell = dataSource.cell(for: message) as? UIView & SelectableView else { return }
+    // MARK: - MessageActionResponder
 
-        wants(toPerform: action, for: message, view: cell)
+    public func perform(action: MessageAction, for message: ZMConversationMessage!, view: UIView) {
+
+        let shouldDismissModal = action != .delete && action != .copy
+
+        if messagePresenter.modalTargetController?.presentedViewController != nil &&
+            shouldDismissModal {
+            messagePresenter.modalTargetController?.dismiss(animated: true) {
+                self.messageAction(actionId: action,
+                                   for: message,
+                                   view: view)
+            }
+        } else {
+            messageAction(actionId: action,
+                          for: message,
+                          view: view)
+        }
     }
 
     func conversationMessageWantsToOpenUserDetails(_ cell: UIView, user: UserType, sourceView: UIView, frame: CGRect) {
@@ -47,3 +61,123 @@ extension ConversationContentViewController: ConversationMessageCellDelegate {
 
 }
 
+extension ConversationContentViewController {
+    func openSketch(for message: ZMConversationMessage, in editMode: CanvasViewControllerEditMode) {
+        let canvasViewController = CanvasViewController()
+        if let imageData = message.imageMessageData?.imageData {
+            canvasViewController.sketchImage = UIImage(data: imageData)
+        }
+        canvasViewController.delegate = self
+        canvasViewController.title = message.conversation?.displayName.uppercased()
+        canvasViewController.select(editMode: editMode, animated: false)
+
+        present(canvasViewController.wrapInNavigationController(), animated: true)
+    }
+
+
+    private func messageAction(actionId: MessageAction,
+                               for message: ZMConversationMessage,
+                               view: UIView) {
+        guard let session = ZMUserSession.shared() else { return }
+
+        switch actionId {
+        case .cancel:
+            session.enqueueChanges({
+                message.fileMessageData?.cancelTransfer()
+            })
+        case .resend:
+            session.enqueueChanges({
+                message.resend()
+            })
+        case .delete:
+            assert(message.canBeDeleted)
+
+            deletionDialogPresenter = DeletionDialogPresenter(sourceViewController: presentedViewController ?? self)
+            deletionDialogPresenter.presentDeletionAlertController(forMessage: message, source: view) { deleted in
+                if deleted {
+                    self.presentedViewController?.dismiss(animated: true)
+                }
+            }
+        case .present:
+            dataSource?.selectedMessage = message
+            presentDetails(for: message)
+        case .save:
+            if Message.isImage(message) {
+                saveImage(from: message, view: view)
+            } else {
+                dataSource?.selectedMessage = message
+
+                let targetView: UIView
+
+                if let selectableView = view as? SelectableView {
+                    targetView = selectableView.selectionView
+                } else {
+                    targetView = view
+                }
+
+                if let saveController = UIActivityViewController(message: message, from: targetView) {
+                    present(saveController, animated: true)
+                }
+            }
+        case .edit:
+            dataSource?.editingMessage = message
+            delegate.conversationContentViewController(self, didTriggerEditing: message)
+        case .sketchDraw:
+            openSketch(for: message, in: .draw)
+        case .sketchEmoji:
+            openSketch(for: message, in: .emoji)
+        case .sketchText:
+            // Not implemented yet
+            break
+        case .like:
+            // The new liked state, the value is flipped
+            let updatedLikedState = !Message.isLikedMessage(message)
+
+            guard let indexPath = dataSource?.indexPath(for: message) else { return }
+
+            let selectedMessage = dataSource?.selectedMessage
+
+            session.performChanges({
+                Message.setLikedMessage(message, liked: updatedLikedState)
+            })
+
+            if updatedLikedState {
+                // Deselect if necessary to show list of likers
+                if selectedMessage == message {
+                    willSelectRow(at: indexPath, tableView: tableView)
+                }
+            } else {
+                // Select if necessary to prevent message from collapsing
+                if !(selectedMessage == message) && !Message.hasReactions(message) {
+                    willSelectRow(at: indexPath, tableView: tableView)
+
+                    tableView.selectRow(at: indexPath, animated: false, scrollPosition: .none)
+                }
+            }
+        case .forward:
+            showForwardFor(message: message, fromCell: view)
+        case .showInConversation:
+            scroll(to: message) { cell in
+                self.dataSource?.highlight(message: message)
+            }
+        case .copy:
+            Message.copy(message, in: UIPasteboard.general)
+        case .download:
+            session.enqueueChanges({
+                message.fileMessageData?.requestFileDownload()
+            })
+        case .reply:
+            delegate.conversationContentViewController(self, didTriggerReplyingTo: message)
+
+        case .openQuote:
+            if let quote = message.textMessageData?.quote {
+                scroll(to: quote) { cell in
+                    self.dataSource?.highlight(message: quote)
+                }
+            }
+        case .openDetails:
+            let detailsViewController = MessageDetailsViewController(message: message)
+            parent?.present(detailsViewController, animated: true)
+        }
+    }
+}
